@@ -14,66 +14,65 @@
   "dependencies": {
     "@actions/core": "^1.10.1",
     "@actions/github": "^6.0.0",
-    "@angular/animations": "19.2.4",
-    "@angular/cdk": "19.2.7",
-    "@angular/common": "19.2.4",
-    "@angular/compiler": "19.2.4",
-    "@angular/core": "19.2.4",
-    "@angular/forms": "19.2.4",
-    "@angular/material": "19.2.7",
-    "@angular/platform-browser": "19.2.4",
-    "@angular/platform-browser-dynamic": "19.2.4",
-    "@angular/router": "19.2.4",
+    "@angular/animations": "20.0.6",
+    "@angular/cdk": "20.0.5",
+    "@angular/common": "20.0.6",
+    "@angular/compiler": "20.0.6",
+    "@angular/core": "20.0.6",
+    "@angular/forms": "20.0.6",
+    "@angular/material": "20.0.5",
+    "@angular/platform-browser": "20.0.6",
+    "@angular/platform-browser-dynamic": "20.0.6",
+    "@angular/platform-server": "20.0.6",
+    "@angular/router": "20.0.6",
+    "@angular/ssr": "20.0.5",
     "@ngneat/falso": "7.2.0",
-    "@ngrx/component": "19.0.1",
     "@ngrx/component-store": "19.0.1",
-    "@ngrx/effects": "19.0.1",
-    "@ngrx/entity": "19.0.1",
     "@ngrx/operators": "19.0.1",
-    "@ngrx/router-store": "19.0.1",
-    "@ngrx/store": "19.0.1",
-    "@nx/angular": "20.6.4",
+    "@nx/angular": "21.2.1",
     "@swc/helpers": "0.5.12",
-    "@tanstack/angular-query-experimental": "5.62.3",
+    "@tanstack/angular-query-experimental": "5.81.5",
+    "express": "^4.21.2",
     "rxjs": "7.8.1",
     "tailwindcss": "3.4.3",
     "tslib": "^2.3.0",
     "zone.js": "0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "19.2.5",
-    "@angular-devkit/core": "19.2.5",
-    "@angular-devkit/schematics": "19.2.5",
-    "@angular-eslint/eslint-plugin": "19.3.0",
-    "@angular-eslint/eslint-plugin-template": "19.3.0",
-    "@angular-eslint/template-parser": "19.3.0",
-    "@angular/cli": "~19.2.0",
-    "@angular/compiler-cli": "19.2.4",
-    "@angular/language-service": "19.2.4",
+    "@angular-devkit/build-angular": "20.0.5",
+    "@angular-devkit/core": "20.0.5",
+    "@angular-devkit/schematics": "20.0.5",
+    "@angular-eslint/eslint-plugin": "20.1.1",
+    "@angular-eslint/eslint-plugin-template": "20.1.1",
+    "@angular-eslint/template-parser": "20.1.1",
+    "@angular/build": "20.0.5",
+    "@angular/cli": "20.0.5",
+    "@angular/compiler-cli": "20.0.6",
+    "@angular/language-service": "20.0.6",
     "@commitlint/cli": "^17.2.0",
     "@commitlint/config-conventional": "^17.2.0",
     "@cypress/webpack-dev-server": "3.8.0",
     "@hirez_io/observer-spy": "^2.2.0",
     "@ngrx/schematics": "19.0.1",
-    "@nx/cypress": "20.6.4",
-    "@nx/devkit": "20.6.4",
-    "@nx/eslint": "20.6.4",
-    "@nx/eslint-plugin": "20.6.4",
-    "@nx/jest": "20.6.4",
-    "@nx/js": "20.6.4",
-    "@nx/plugin": "20.6.4",
-    "@nx/web": "20.6.4",
-    "@nx/workspace": "20.6.4",
-    "@schematics/angular": "19.2.5",
+    "@nx/cypress": "21.2.1",
+    "@nx/devkit": "21.2.1",
+    "@nx/eslint": "21.2.1",
+    "@nx/eslint-plugin": "21.2.1",
+    "@nx/jest": "21.2.1",
+    "@nx/js": "21.2.1",
+    "@nx/plugin": "21.2.1",
+    "@nx/web": "21.2.1",
+    "@nx/workspace": "21.2.1",
+    "@schematics/angular": "20.0.4",
     "@swc-node/register": "1.9.2",
-    "@swc/cli": "0.5.2",
+    "@swc/cli": "0.6.0",
     "@swc/core": "1.10.0",
     "@tailwindcss/forms": "0.5.9",
-    "@tanstack/angular-query-devtools-experimental": "5.38.0",
-    "@testing-library/angular": "16.0.0",
+    "@testing-library/angular": "17.4.0",
     "@testing-library/cypress": "10.0.3",
-    "@testing-library/jest-dom": "6.4.5",
-    "@testing-library/user-event": "14.5.2",
+    "@testing-library/jest-dom": "6.6.3",
+    "@testing-library/user-event": "14.6.1",
+    "@types/express": "^4.17.21",
     "@types/jest": "29.5.13",
     "@types/node": "18.16.9",
     "@typescript-eslint/eslint-plugin": "7.16.1",
@@ -102,15 +101,22 @@
     "prettier-plugin-organize-imports": "3.2.4",
     "prettier-plugin-tailwindcss": "0.5.14",
     "source-map-explorer": "^2.5.3",
-    "testing-library-selector": "^0.3.1",
     "ts-jest": "29.1.0",
     "ts-node": "10.9.1",
-    "typescript": "5.7.3"
+    "typescript": "5.8.3"
   },
   "overrides": {
     "chokidar": "4.0.1",
     "@nx/cypress": {
       "cypress": "14.0.2"
+    },
+    "@ngrx/component-store": {
+      "@angular/common": "$@angular/common",
+      "@angular/core": "$@angular/core"
+    },
+    "@ngrx/operators": {
+      "@angular/common": "$@angular/common",
+      "@angular/core": "$@angular/core"
     }
   },
   "lint-staged": {
