@@ -1,11 +1,11 @@
 /* eslint-disable @angular-eslint/component-selector */
-import { Component, Input } from '@angular/core';
+import { Component, input } from '@angular/core';
+
 @Component({
   selector: 'nav-button',
-  standalone: true,
   template: `
-    <a [href]="href">
-      <ng-content></ng-content>
+    <a [href]="href()">
+      <ng-content />
     </a>
   `,
   host: {
@@ -13,5 +13,5 @@ import { Component, Input } from '@angular/core';
   },
 })
 export class NavButtonComponent {
-  @Input() href = '';
+  href = input('');
 }
