@@ -1,5 +1,5 @@
-import { AsyncPipe, NgFor, NgIf } from '@angular/common';
-import { Component, Input, inject } from '@angular/core';
+import { Component, inject, input } from '@angular/core';
+import { toSignal } from '@angular/core/rxjs-interop';
 import { RouterLink, RouterLinkActive } from '@angular/router';
 import { FakeServiceService } from './fake.service';
 
@@ -10,16 +10,16 @@ interface MenuItem {
 
 @Component({
   selector: 'app-nav',
-  imports: [RouterLink, RouterLinkActive, NgFor],
+  imports: [RouterLink, RouterLinkActive],
   template: `
-    <ng-container *ngFor="let menu of menus">
+    @for (menu of menus(); track menu.path) {
       <a
         class="rounded-md border px-4 py-2"
         [routerLink]="menu.path"
         routerLinkActive="isSelected">
         {{ menu.name }}
       </a>
-    </ng-container>
+    }
   `,
   styles: [
     `
@@ -33,28 +33,26 @@ interface MenuItem {
   },
 })
 export class NavigationComponent {
-  @Input() menus!: MenuItem[];
+  menus = input.required<MenuItem[]>();
 }
 
 @Component({
-  imports: [NavigationComponent, NgIf, AsyncPipe],
+  imports: [NavigationComponent],
   template: `
-    <ng-container *ngIf="info$ | async as info">
-      <ng-container *ngIf="info !== null; else noInfo">
-        <app-nav [menus]="getMenu(info)" />
-      </ng-container>
-    </ng-container>
-
-    <ng-template #noInfo>
+    @if (info() !== null) {
+      <app-nav [menus]="getMenu(info()!)" />
+    } @else {
       <app-nav [menus]="getMenu('')" />
-    </ng-template>
+    }
   `,
   host: {},
 })
 export class MainNavigationComponent {
   private fakeBackend = inject(FakeServiceService);
 
-  readonly info$ = this.fakeBackend.getInfoFromBackend();
+  readonly info = toSignal(this.fakeBackend.getInfoFromBackend(), {
+    initialValue: null,
+  });
 
   getMenu(prop: string) {
     return [
