@@ -1,31 +1,30 @@
-import { CommonModule } from '@angular/common';
+import { NgTemplateOutlet } from '@angular/common';
 import {
   ChangeDetectionStrategy,
   Component,
-  ContentChild,
-  Input,
+  contentChild,
+  input,
   TemplateRef,
 } from '@angular/core';
 
 @Component({
   selector: 'list',
-  imports: [CommonModule],
   template: `
-    <div *ngFor="let item of list; index as i">
+    @for (item of list(); track $index) {
       <ng-container
         *ngTemplateOutlet="
-          listTemplateRef || emptyRef;
-          context: { $implicit: item, appList: item, index: i }
-        "></ng-container>
-    </div>
+          listTemplateRef() || emptyRef;
+          context: { $implicit: item, appList: item, index: $index }
+        " />
+    }
 
     <ng-template #emptyRef>No Template</ng-template>
   `,
   changeDetection: ChangeDetectionStrategy.OnPush,
+  imports: [NgTemplateOutlet],
 })
 export class ListComponent<TItem extends object> {
-  @Input() list!: TItem[];
+  list = input.required<TItem[]>();
 
-  @ContentChild('listRef', { read: TemplateRef })
-  listTemplateRef!: TemplateRef<unknown>;
+  listTemplateRef = contentChild('listRef', { read: TemplateRef });
 }
