@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, inject } from '@angular/core';
 import { RouterLink } from '@angular/router';
 import { ButtonComponent } from './button.component';
 import { InformationComponent } from './information.component';
@@ -28,7 +28,7 @@ import { UserStore } from './user.store';
       <button app-button (click)="everyone()">Everyone</button>
     </header>
 
-    <app-information></app-information>
+    <app-information />
 
     <button app-button class="mt-10" routerLink="enter">
       Enter application
@@ -36,7 +36,7 @@ import { UserStore } from './user.store';
   `,
 })
 export class LoginComponent {
-  constructor(private userStore: UserStore) {}
+  private readonly userStore = inject(UserStore);
 
   admin() {
     this.userStore.add(admin);
