@@ -1,14 +1,12 @@
-import { NgFor } from '@angular/common';
 import { Component } from '@angular/core';
 
 @Component({
-  imports: [NgFor],
   selector: 'app-root',
   template: `
-    <div *ngFor="let person of persons; let index = index; let isFirst = first">
-      {{ showName(person.name, index) }}
-      {{ isAllowed(person.age, isFirst) }}
-    </div>
+    @for (person of persons; track person.name) {
+      {{ showName(person.name, $index) }}
+      {{ isAllowed(person.age, $first) }}
+    }
   `,
 })
 export class AppComponent {
