@@ -1,6 +1,6 @@
-import { Component, Input } from '@angular/core';
+import { Component, input } from '@angular/core';
 
-import { CommonModule } from '@angular/common';
+import { TitleCasePipe } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { MatChipsModule } from '@angular/material/chips';
 import { MatFormFieldModule } from '@angular/material/form-field';
@@ -18,16 +18,16 @@ const fibonacci = (num: number): number => {
 @Component({
   selector: 'app-person-list',
   imports: [
-    CommonModule,
     FormsModule,
     MatListModule,
     MatFormFieldModule,
     MatInputModule,
     MatChipsModule,
+    TitleCasePipe,
   ],
   template: `
     <h1 class="text-center font-semibold" title="Title">
-      {{ title | titlecase }}
+      {{ title() | titlecase }}
     </h1>
 
     <mat-form-field class="w-4/5">
@@ -39,12 +39,14 @@ const fibonacci = (num: number): number => {
     </mat-form-field>
 
     <mat-list class="flex w-full">
-      <mat-list-item *ngFor="let person of persons">
-        <div MatListItemLine class="flex justify-between">
-          <h3>{{ person.name }}</h3>
-          <mat-chip>{{ calculate(person.fib) }}</mat-chip>
-        </div>
-      </mat-list-item>
+      @for (person of persons(); track person.name) {
+        <mat-list-item>
+          <div class="flex justify-between">
+            <h3>{{ person.name }}</h3>
+            <mat-chip>{{ calculate(person.fib) }}</mat-chip>
+          </div>
+        </mat-list-item>
+      }
     </mat-list>
   `,
   host: {
@@ -52,8 +54,8 @@ const fibonacci = (num: number): number => {
   },
 })
 export class PersonListComponent {
-  @Input() persons: Person[] = [];
-  @Input() title = '';
+  persons = input<Person[]>();
+  title = input('');
 
   label = '';
 
