@@ -1,21 +1,17 @@
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
-
-import { NgForTrackByModule } from '@angular-challenges/shared/directives';
-import { CommonModule } from '@angular/common';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 import { Person } from './person.model';
 
 @Component({
   selector: 'app-person-list',
-  imports: [CommonModule, NgForTrackByModule],
   template: `
     <div class="relative h-[300px] overflow-hidden">
       <div class="absolute inset-0 overflow-scroll">
-        <div
-          *ngFor="let person of persons; trackByProp: 'email'"
-          class="flex h-9 items-center justify-between border-b">
-          <h3>{{ person.name }}</h3>
-          <p>{{ person.email }}</p>
-        </div>
+        @for (person of persons(); track person.email) {
+          <div class="flex h-9 items-center justify-between border-b">
+            <h3>{{ person.name }}</h3>
+            <p>{{ person.email }}</p>
+          </div>
+        }
       </div>
     </div>
   `,
@@ -25,5 +21,5 @@ import { Person } from './person.model';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class PersonListComponent {
-  @Input() persons: Person[] = [];
+  persons = input<Person[]>();
 }
