@@ -1,4 +1,3 @@
-import { NgFor } from '@angular/common';
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MAT_DIALOG_DATA, MatDialogModule } from '@angular/material/dialog';
@@ -8,16 +7,18 @@ import { MAT_DIALOG_DATA, MatDialogModule } from '@angular/material/dialog';
     <h1 mat-dialog-title>Show all Topics</h1>
     <div mat-dialog-content>
       <ul>
-        <li *ngFor="let topic of data.topics">
-          {{ topic }}
-        </li>
+        @for (topic of data.topics; track $index) {
+          <li>
+            {{ topic }}
+          </li>
+        }
       </ul>
     </div>
     <div mat-dialog-actions>
       <button mat-button mat-dialog-close>Close</button>
     </div>
   `,
-  imports: [MatDialogModule, MatButtonModule, NgFor],
+  imports: [MatDialogModule, MatButtonModule],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class TopicModalComponent {
