@@ -1,13 +1,18 @@
-import { NgFor, NgIf } from '@angular/common';
-import { Component, OnInit, inject } from '@angular/core';
+import { AsyncPipe } from '@angular/common';
+import { Component, inject, OnInit } from '@angular/core';
 import { FormControl, ReactiveFormsModule } from '@angular/forms';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
 import { RouterLinkWithHref } from '@angular/router';
-import { LetDirective } from '@ngrx/component';
 import { provideComponentStore } from '@ngrx/component-store';
-import { debounceTime, distinctUntilChanged, skipWhile, tap } from 'rxjs';
+import {
+  debounceTime,
+  distinctUntilChanged,
+  Observable,
+  skipWhile,
+  tap,
+} from 'rxjs';
 import { Photo } from '../photo.model';
 import { PhotoStore } from './photos.store';
 
@@ -17,11 +22,9 @@ import { PhotoStore } from './photos.store';
     ReactiveFormsModule,
     MatFormFieldModule,
     MatProgressBarModule,
-    NgIf,
-    NgFor,
     MatInputModule,
-    LetDirective,
     RouterLinkWithHref,
+    AsyncPipe,
   ],
   template: `
     <h2 class="mb-2 text-xl">Photos</h2>
@@ -35,7 +38,7 @@ import { PhotoStore } from './photos.store';
         placeholder="find a photo" />
     </mat-form-field>
 
-    <ng-container *ngrxLet="vm$ as vm">
+    @let vm = vm$ | async;
       <section class="flex flex-col">
         <section class="flex items-center gap-3">
           <button
@@ -54,30 +57,33 @@ import { PhotoStore } from './photos.store';
           </button>
           Page :{{ vm.page }} / {{ vm.pages }}
         </section>
-        <mat-progress-bar
-          mode="query"
-          *ngIf="vm.loading"
-          class="mt-5"></mat-progress-bar>
-        <ul
-          class="flex flex-wrap gap-4"
-          *ngIf="vm.photos && vm.photos.length > 0; else noPhoto">
-          <li *ngFor="let photo of vm.photos; trackBy: trackById">
-            <a routerLink="detail" [queryParams]="{ photo: encode(photo) }">
-              <img
-                src="{{ photo.url_q }}"
-                alt="{{ photo.title }}"
-                class="image" />
-            </a>
-          </li>
-        </ul>
-        <ng-template #noPhoto>
+        @if (vm.loading) {
+          <mat-progress-bar mode="query" class="mt-5"></mat-progress-bar>
+        }
+        @if (vm.photos && vm.photos.length > 0) {
+          <ul class="flex flex-wrap gap-4">
+            @for (
+              photo of vm.photos;
+              track photo.id;
+              let i = $index
+            ) {
+              <li>
+                <a routerLink="detail" [queryParams]="{ photo: encode(photo) }">
+                  <img
+                    src="{{ photo.url_q }}"
+                    alt="{{ photo.title }}"
+                    class="image" />
+                </a>
+              </li>
+            }
+          </ul>
+        } @else {
           <div>No Photos found. Type a search word.</div>
-        </ng-template>
+        }
         <footer class="text-red-500">
           {{ vm.error }}
         </footer>
       </section>
-    </ng-container>
   `,
   providers: [provideComponentStore(PhotoStore)],
   host: {
@@ -86,7 +92,15 @@ import { PhotoStore } from './photos.store';
 })
 export default class PhotosComponent implements OnInit {
   store = inject(PhotoStore);
-  readonly vm$ = this.store.vm$.pipe(
+  readonly vm$: Observable<{
+    photos: Photo[];
+    search: string;
+    page: number;
+    pages: number;
+    endOfPage: boolean;
+    loading: boolean;
+    error: unknown;
+  }> = this.store.vm$.pipe(
     tap(({ search }) => {
       if (!this.formInit) {
         this.search.setValue(search);
@@ -106,10 +120,6 @@ export default class PhotosComponent implements OnInit {
         distinctUntilChanged(),
       ),
     );
-  }
-
-  trackById(index: number, photo: Photo) {
-    return photo.id;
   }
 
   encode(photo: Photo) {
