@@ -1,4 +1,4 @@
-import { AsyncPipe, JsonPipe, NgFor } from '@angular/common';
+import { AsyncPipe } from '@angular/common';
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { map } from 'rxjs';
@@ -6,15 +6,14 @@ import { availableBooks } from './book.model';
 
 @Component({
   selector: 'app-shelf',
-  imports: [AsyncPipe, JsonPipe, NgFor],
+  imports: [AsyncPipe],
   template: `
     <ul>
-      <li *ngFor="let book of books | async">
-        Borrowed Book: {{ book.name }} by {{ book.author }}
-      </li>
+      @for (book of books | async; track book) {
+        <li>Borrowed Book: {{ book.name }} by {{ book.author }}</li>
+      }
     </ul>
   `,
-  styles: [],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export default class ShelfComponent {
