@@ -1,10 +1,8 @@
-import { Component, Input, OnInit, Output, signal } from '@angular/core';
-import { toObservable } from '@angular/core/rxjs-interop';
+import { Component, input, OnInit, output, signal } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSliderModule } from '@angular/material/slider';
-import { skip } from 'rxjs';
 
 @Component({
   selector: 'app-slider',
@@ -15,19 +13,19 @@ import { skip } from 'rxjs';
           <button id="minusButton" mat-mini-fab (click)="back()">
             <mat-icon>arrow_back_ios</mat-icon>
           </button>
-          {{ minValue }}
+          {{ minValue() }}
           <mat-slider
             class="m-4"
-            [max]="maxValue"
-            [min]="minValue"
-            [disabled]="disabled"
-            [step]="step">
+            [max]="maxValue()"
+            [min]="minValue()"
+            [disabled]="disabled()"
+            [step]="step()">
             <input
               matSliderThumb
               [value]="value()"
               (valueChange)="value.set($event)" />
           </mat-slider>
-          {{ maxValue }}
+          {{ maxValue() }}
           <button id="plusButton" mat-mini-fab (click)="forward()">
             <mat-icon>arrow_forward_ios</mat-icon>
           </button>
@@ -51,27 +49,27 @@ import { skip } from 'rxjs';
   imports: [MatCardModule, MatSliderModule, MatIconModule, FormsModule],
 })
 export class SliderComponent implements OnInit {
-  @Input() step = 1;
-  @Input() minValue = 0;
-  @Input() maxValue = 100;
-  @Input() disabled = false;
+  step = input(1);
+  minValue = input(0);
+  maxValue = input(100);
+  disabled = input(false);
 
   value = signal(0);
-  @Output() valueChange = toObservable(this.value).pipe(skip(1));
+  valueChange = output<number>();
 
   ngOnInit(): void {
-    this.value.set(this.minValue);
+    this.value.set(this.minValue());
   }
 
   back() {
-    if (this.value() - this.step >= this.minValue) {
-      this.value.update((v) => v - this.step);
+    if (this.value() - this.step() >= this.minValue()) {
+      this.value.update((v) => v - this.step());
     }
   }
 
   forward() {
-    if (this.value() + this.step <= this.maxValue) {
-      this.value.update((v) => v + this.step);
+    if (this.value() + this.step() <= this.maxValue()) {
+      this.value.update((v) => v + this.step());
     }
   }
 }
