@@ -1,10 +1,9 @@
-import { NgFor, NgIf } from '@angular/common';
-import { Component, OnInit, inject } from '@angular/core';
+import { AsyncPipe } from '@angular/common';
+import { Component, inject, OnInit } from '@angular/core';
 import { FormControl, ReactiveFormsModule } from '@angular/forms';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
-import { LetDirective } from '@ngrx/component';
 import { provideComponentStore } from '@ngrx/component-store';
 import { TicketStore } from './ticket.store';
 import { AddComponent } from './ui/add.component';
@@ -18,10 +17,8 @@ import { RowComponent } from './ui/row.component';
     RowComponent,
     MatFormFieldModule,
     MatProgressBarModule,
-    NgIf,
-    NgFor,
     MatInputModule,
-    LetDirective,
+    AsyncPipe,
   ],
   template: `
     <h2 class="mb-2 text-xl">Tickets</h2>
@@ -35,27 +32,27 @@ import { RowComponent } from './ui/row.component';
         placeholder="write an article" />
     </mat-form-field>
 
-    <ng-container *ngrxLet="vm$ as vm">
+    @if (vm$ | async; as vm) {
       <app-add
         [loading]="vm.loading"
         (addTicket)="ticketStore.addTicket($event)"></app-add>
 
-      <mat-progress-bar
-        mode="query"
-        *ngIf="vm.loading"
-        class="mt-5"></mat-progress-bar>
+      @if (vm.loading) {
+        <mat-progress-bar mode="query" class="mt-5"></mat-progress-bar>
+      }
       <ul class="flex max-w-3xl flex-col gap-4">
-        <app-row
-          *ngFor="let t of vm.tickets"
-          [ticket]="t"
-          [users]="vm.users"
-          (assign)="ticketStore.assignTicket($event)"
-          (closeTicket)="ticketStore.done($event)"></app-row>
+        @for (ticket of vm.tickets; track ticket.id) {
+          <app-row
+            [ticket]="ticket"
+            [users]="vm.users"
+            (assign)="ticketStore.assignTicket($event)"
+            (closeTicket)="ticketStore.done($event)"></app-row>
+        }
       </ul>
       <footer class="text-red-500">
         {{ vm.error }}
       </footer>
-    </ng-container>
+    }
   `,
   providers: [provideComponentStore(TicketStore)],
   host: {
