@@ -1,5 +1,5 @@
 /* eslint-disable @angular-eslint/component-selector */
-import { Component, Input } from '@angular/core';
+import { Component, computed, input } from '@angular/core';
 import { TextComponent } from './text.component';
 
 export type StaticTextType = 'normal' | 'warning' | 'error';
@@ -8,25 +8,31 @@ export type StaticTextType = 'normal' | 'warning' | 'error';
   selector: 'static-text',
   imports: [TextComponent],
   template: `
-    <text [font]="font" [color]="color">This is a static text</text>
+    <text [font]="font()" [color]="color()">This is a static text</text>
   `,
 })
 export class TextStaticComponent {
-  @Input() set type(type: StaticTextType) {
-    switch (type) {
-      case 'error': {
-        this.font = 30;
-        this.color = 'red';
-        break;
-      }
-      case 'warning': {
-        this.font = 25;
-        this.color = 'orange';
-        break;
-      }
-    }
-  }
+  type = input<StaticTextType>('normal');
 
-  font = 10;
-  color = 'black';
+  font = computed(() => {
+    switch (this.type()) {
+      case 'error':
+        return 30;
+      case 'warning':
+        return 25;
+      default:
+        return 10;
+    }
+  });
+
+  color = computed(() => {
+    switch (this.type()) {
+      case 'error':
+        return 'red';
+      case 'warning':
+        return 'orange';
+      default:
+        return 'black';
+    }
+  });
 }
