@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
+
 @Component({
   selector: 'app-bar',
-  standalone: true,
   template: `
     BarComponent
   `,
