@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
+
 @Component({
   selector: 'app-foo',
-  standalone: true,
   template: `
     Foo Component
   `,
