@@ -1,10 +1,8 @@
-import { CommonModule } from '@angular/common';
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { UserStore } from './user.store';
 
 @Component({
   selector: 'app-information',
-  imports: [CommonModule],
   template: `
     <h2 class="mt-10 text-xl">Information Panel</h2>
     <!-- admin can see everything -->
@@ -18,6 +16,7 @@ import { UserStore } from './user.store';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class InformationComponent {
+  private readonly userStore = inject(UserStore);
+
   user$ = this.userStore.user$;
-  constructor(private userStore: UserStore) {}
 }
