@@ -2,5 +2,7 @@ import { nxComponentTestingPreset } from '@nx/angular/plugins/component-testing'
 import { defineConfig } from 'cypress';
 
 export default defineConfig({
-  component: nxComponentTestingPreset(__filename),
+  component: {
+    ...nxComponentTestingPreset(__filename),
+  },
 });
