@@ -1,12 +1,11 @@
 /* eslint-disable @angular-eslint/directive-selector */
 /* eslint-disable @angular-eslint/no-host-metadata-property */
-import { Directive, WritableSignal, signal } from '@angular/core';
+import { Directive, signal, WritableSignal } from '@angular/core';
 
 export type ButtonState = 'enabled' | 'disabled';
 
 @Directive({
   selector: 'button[btnDisabled]',
-  standalone: true,
   host: {
     '(click)': 'toggleState()',
   },
